split benchmarks definition from program
